rename AssignedContactsGrid to AllAssignedContactsGrid
