+++ packages/admin/src/targetGroups/TargetGroupForm.tsx
@@ -27,7 +27,7 @@ import { FormattedMessage } from "react-intl";
 export { namedOperations as targetGroupFormNamedOperations } from "./TargetGroupForm.gql.generated";
 
 import { AddContactsGridSelect } from "./addContacts/AddContactsGridSelect";
-import { AssignedContactsGrid } from "./allAssignedContacts/AssignedContactsGrid";
+import { AllAssignedContactsGrid } from "./allAssignedContacts/AllAssignedContactsGrid";
 import { createTargetGroupMutation, targetGroupFormQuery, updateTargetGroupMutation } from "./TargetGroupForm.gql";
 import {
     GQLCreateTargetGroupMutation,
@@ -203,7 +203,7 @@ export function TargetGroupForm({ id, scope, additionalFormFields, input2State, 
                                     disablePadding
                                     initiallyExpanded={false}
                                 >
-                                    <AssignedContactsGrid brevoId={data?.targetGroup.brevoId ?? undefined} id={id} scope={scope} />
+                                    <AllAssignedContactsGrid brevoId={data?.targetGroup.brevoId ?? undefined} id={id} scope={scope} />
                                 </FieldSet>
                             </>
                         )}
